fix: recorder hardening — 64-bit handles, LoadTexture coverage, thread guard
Review follow-ups on 58419ed before the replayer consumes the contract:
- FVaCuusGeometryHandle/FVaCuusTextureHandle widened to uint64 (round-trips
  Rml uintptr_t handles; static_asserts added; wrap ensureMsgf on counters)
- New VaCuus.Render.Recorder.LoadTexture test: in-test 4x2 PNG encode,
  happy path (handle, dimensions, lossless RGBA round-trip) + missing-file
  and non-image-content failure branches
- check() narrowing guards on vertex/index/texel counts
- Thread-agnostic single-writer doc + owner-thread ensure between
  BeginFrame and EndFrameAndPublish
- Destructor log tripwire for unpublished pending traffic (log, not
  ensure: the M1 teardown path legitimately drops trailing releases)
- Integration test asserts engine is down before installing the recorder
- EndFrameAndPublish doc notes pre-populated fresh buffers

All 7 VaCuus tests green under -nullrhi.

+++ Source/VaCuusRender/Public/VaCuusCommandBuffer.h
@@ -5,13 +5,14 @@
 #include "CoreMinimal.h"
 
 /**
- * Dense 1-based ids minted by the recorder (index+1); 0 = invalid/none.
+ * Dense 1-based ids minted by the recorder; 0 = invalid/none.
  * Handles are shared with RmlUi verbatim: the recorder returns them as
- * Rml::CompiledGeometryHandle / Rml::TextureHandle and gets them back in
- * draw and release calls.
+ * Rml::CompiledGeometryHandle / Rml::TextureHandle (uintptr_t, 64-bit on all
+ * target platforms) and gets them back in draw and release calls, so the
+ * full 64-bit range round-trips without truncation.
  */
-using FVaCuusGeometryHandle = uint32;
-using FVaCuusTextureHandle = uint32;
+using FVaCuusGeometryHandle = uint64;
+using FVaCuusTextureHandle = uint64;
 
 enum class EVaCuusCommandType : uint8
 {
